added workflow runner, metric enums, formatting
